Artificial Intelligence (AI)/Tutorials/Retrieval Augmented Generation (RAG)/4. Enhanced RAG with Re-ranking/requirements.txt
openai
langchain
langchain-community
chromadb
umap-learn
matplotlib
pypdf
python-dotenv
numpy
pandas
psutil
platformdirs
sentence_transformers